fix: make theme logic into custom hooks and remove theme folder for better file sturcture
