+++ src/hooks/useTheme.tsx
@@ -1,0 +1,22 @@
+import { useEffect, useState } from "react";
+
+export const useTheme = () => {
+	const [theme, setThemeState] = useState<"theme-light" | "dark" | "system">(
+		"theme-light",
+	);
+
+	useEffect(() => {
+		const isDarkMode = document.documentElement.classList.contains("dark");
+		setThemeState(isDarkMode ? "dark" : "theme-light");
+	}, []);
+
+	useEffect(() => {
+		const isDark =
+			theme === "dark" ||
+			(theme === "system" &&
+				window.matchMedia("(prefers-color-scheme: dark)").matches);
+		document.documentElement.classList[isDark ? "add" : "remove"]("dark");
+	}, [theme]);
+
+	return [theme, setThemeState] as const;
+};
